CompiledExpression into its own file
