+++ rust/routee-compass/src/plugin/output/default/eval/mod.rs
@@ -1,10 +1,12 @@
 mod builder;
+mod compiled_expression;
 mod config;
+mod operation;
 mod plugin;
 
-mod operation;
 pub mod ops;
 pub use builder::EvalOutputPluginBuilder;
+pub use compiled_expression::CompiledExpression;
 pub use config::{EvalOutputPluginConfig, ExpressionConfig, NotANumberBehavior, OnFailureBehavior};
 pub use operation::Operation;
 pub use plugin::EvalOutputPlugin;
