+++ rust/routee-compass/src/plugin/output/default/eval/ops.rs
@@ -6,7 +6,10 @@ use serde_json::Value;
 use serde_json_path::JsonPath;
 
 use crate::plugin::output::{
-    default::eval::{config::ExpressionConfig, NotANumberBehavior, Operation},
+    default::eval::{
+        compiled_expression::CompiledExpression, config::ExpressionConfig, NotANumberBehavior,
+        Operation,
+    },
     OutputPluginError,
 };
 
@@ -17,21 +20,21 @@ pub enum PathSegment {
     Index(usize),
 }
 
-/// One expression with its inputs pre-parsed into [`JsonPath`]s and its output
-/// path pre-parsed into [`PathSegment`]s so the per-query hot path is as cheap
-/// as possible.
-pub struct CompiledExpression {
-    /// `(variable_name, compiled JSONPath)` pairs for each input binding.
-    pub inputs: Vec<(String, JsonPath)>,
-    /// Raw expression string retained for error messages.
-    pub expr: String,
-    /// Pre-parsed output path segments.
-    pub output_segments: Vec<PathSegment>,
-    /// Memory arena that backs the pre-compiled bytecode.
-    pub slab: fasteval::Slab,
-    /// Pre-compiled bytecode produced by `fasteval`'s compiler.
-    pub compiled: fasteval::Instruction,
-}
+// /// One expression with its inputs pre-parsed into [`JsonPath`]s and its output
+// /// path pre-parsed into [`PathSegment`]s so the per-query hot path is as cheap
+// /// as possible.
+// pub struct CompiledExpression {
+//     /// `(variable_name, compiled JSONPath)` pairs for each input binding.
+//     pub inputs: Vec<(String, JsonPath)>,
+//     /// Raw expression string retained for error messages.
+//     pub expr: String,
+//     /// Pre-parsed output path segments.
+//     pub output_segments: Vec<PathSegment>,
+//     /// Memory arena that backs the pre-compiled bytecode.
+//     pub slab: fasteval::Slab,
+//     /// Pre-compiled bytecode produced by `fasteval`'s compiler.
+//     pub compiled: fasteval::Instruction,
+// }
 
 /// Compiled form of [`OnFailureBehavior`] — the `Record` variant's path is
 /// pre-parsed once at plugin construction time.
@@ -48,188 +51,122 @@ pub enum CompiledOnFailure {
 // Config → CompiledExpression
 // ---------------------------------------------------------------------------
 
-pub fn compile_expression(
-    conf: ExpressionConfig,
-) -> Result<CompiledExpression, crate::plugin::PluginError> {
-    let inputs = conf
-        .inputs
-        .into_iter()
-        .map(|(name, path_str)| {
-            let path = JsonPath::parse(&path_str).map_err(|e| {
-                crate::plugin::PluginError::BuildFailed(format!(
-                    "invalid JSONPath '{path_str}' for input '{name}': {e}"
-                ))
-            })?;
-            Ok((name, path))
-        })
-        .collect::<Result<Vec<_>, crate::plugin::PluginError>>()?;
+// pub fn compile_expression(
+//     conf: ExpressionConfig,
+// ) -> Result<CompiledExpression, crate::plugin::PluginError> {
+//     let inputs = conf
+//         .inputs
+//         .into_iter()
+//         .map(|(name, path_str)| {
+//             let path = JsonPath::parse(&path_str).map_err(|e| {
+//                 crate::plugin::PluginError::BuildFailed(format!(
+//                     "invalid JSONPath '{path_str}' for input '{name}': {e}"
+//                 ))
+//             })?;
+//             Ok((name, path))
+//         })
+//         .collect::<Result<Vec<_>, crate::plugin::PluginError>>()?;
 
-    let output_segments = parse_path_segments(&conf.output).map_err(|e| {
-        crate::plugin::PluginError::BuildFailed(format!(
-            "invalid output path '{}': {e}",
-            conf.output
-        ))
-    })?;
+//     let output_segments = parse_path_segments(&conf.output).map_err(|e| {
+//         crate::plugin::PluginError::BuildFailed(format!(
+//             "invalid output path '{}': {e}",
+//             conf.output
+//         ))
+//     })?;
 
-    // Parse and compile the fasteval expression once so that eval_and_write can
-    // skip the parse step on every row.
-    let mut slab = fasteval::Slab::new();
-    let parsed = fasteval::Parser::new()
-        .parse(&conf.expr, &mut slab.ps)
-        .map_err(|e| {
-            crate::plugin::PluginError::BuildFailed(format!(
-                "failed to parse expression '{}': {e}",
-                conf.expr
-            ))
-        })?;
-    let compiled = parsed.from(&slab.ps).compile(&slab.ps, &mut slab.cs);
+//     // Parse and compile the fasteval expression once so that eval_and_write can
+//     // skip the parse step on every row.
+//     let mut slab = fasteval::Slab::new();
+//     let parsed = fasteval::Parser::new()
+//         .parse(&conf.expr, &mut slab.ps)
+//         .map_err(|e| {
+//             crate::plugin::PluginError::BuildFailed(format!(
+//                 "failed to parse expression '{}': {e}",
+//                 conf.expr
+//             ))
+//         })?;
+//     let compiled = parsed.from(&slab.ps).compile(&slab.ps, &mut slab.cs);
 
-    Ok(CompiledExpression {
-        inputs,
-        expr: conf.expr,
-        output_segments,
-        slab,
-        compiled,
-    })
-}
+//     Ok(CompiledExpression {
+//         inputs,
+//         expr: conf.expr,
+//         output_segments,
+//         slab,
+//         compiled,
+//     })
+// }
 
 // ---------------------------------------------------------------------------
 // Path helpers
 // ---------------------------------------------------------------------------
 
-/// Parse a JSONPath-style string into segments for writing.
-///
-/// Supports dot notation (`$.a.b`) and bracket array indices (`$.a[0].b`).
-/// The leading `$.` or `$` prefix is stripped before parsing.
-pub fn parse_path_segments(path: &str) -> Result<Vec<PathSegment>, OutputPluginError> {
-    let stripped = path
-        .strip_prefix("$.")
-        .or_else(|| path.strip_prefix('$'))
-        .unwrap_or(path);
+// /// Parse a JSONPath-style string into segments for writing.
+// ///
+// /// Supports dot notation (`$.a.b`) and bracket array indices (`$.a[0].b`).
+// /// The leading `$.` or `$` prefix is stripped before parsing.
+// pub fn parse_path_segments(path: &str) -> Result<Vec<PathSegment>, OutputPluginError> {
+//     let stripped = path
+//         .strip_prefix("$.")
+//         .or_else(|| path.strip_prefix('$'))
+//         .unwrap_or(path);
 
-    if stripped.is_empty() {
-        return Err(OutputPluginError::OutputPluginFailed(
-            "output path must not be empty after '$'".to_string(),
-        ));
-    }
+//     if stripped.is_empty() {
+//         return Err(OutputPluginError::OutputPluginFailed(
+//             "output path must not be empty after '$'".to_string(),
+//         ));
+//     }
 
-    let mut segments = Vec::new();
-    let mut current_key = String::new();
-    let mut chars = stripped.chars().peekable();
+//     let mut segments = Vec::new();
+//     let mut current_key = String::new();
+//     let mut chars = stripped.chars().peekable();
 
-    while let Some(c) = chars.next() {
-        match c {
-            '.' => {
-                if !current_key.is_empty() {
-                    segments.push(PathSegment::Key(std::mem::take(&mut current_key)));
-                }
-            }
-            '[' => {
-                if !current_key.is_empty() {
-                    segments.push(PathSegment::Key(std::mem::take(&mut current_key)));
-                }
-                let mut bracket_content = String::new();
-                for ic in chars.by_ref() {
-                    if ic == ']' {
-                        break;
-                    }
-                    bracket_content.push(ic);
-                }
-                // Quoted string key: ['foo'] or ["foo"]
-                let segment = if (bracket_content.starts_with('\'')
-                    && bracket_content.ends_with('\''))
-                    || (bracket_content.starts_with('"') && bracket_content.ends_with('"'))
-                {
-                    let key = bracket_content[1..bracket_content.len() - 1].to_string();
-                    PathSegment::Key(key)
-                } else {
-                    // Numeric array index: [0]
-                    let idx: usize = bracket_content.parse().map_err(|_| {
-                        OutputPluginError::OutputPluginFailed(format!(
-                            "invalid bracket segment '[{bracket_content}]' in path '{path}': \
-                            expected an integer index or a quoted string key"
-                        ))
-                    })?;
-                    PathSegment::Index(idx)
-                };
-                segments.push(segment);
-            }
-            other => current_key.push(other),
-        }
-    }
+//     while let Some(c) = chars.next() {
+//         match c {
+//             '.' => {
+//                 if !current_key.is_empty() {
+//                     segments.push(PathSegment::Key(std::mem::take(&mut current_key)));
+//                 }
+//             }
+//             '[' => {
+//                 if !current_key.is_empty() {
+//                     segments.push(PathSegment::Key(std::mem::take(&mut current_key)));
+//                 }
+//                 let mut bracket_content = String::new();
+//                 for ic in chars.by_ref() {
+//                     if ic == ']' {
+//                         break;
+//                     }
+//                     bracket_content.push(ic);
+//                 }
+//                 // Quoted string key: ['foo'] or ["foo"]
+//                 let segment = if (bracket_content.starts_with('\'')
+//                     && bracket_content.ends_with('\''))
+//                     || (bracket_content.starts_with('"') && bracket_content.ends_with('"'))
+//                 {
+//                     let key = bracket_content[1..bracket_content.len() - 1].to_string();
+//                     PathSegment::Key(key)
+//                 } else {
+//                     // Numeric array index: [0]
+//                     let idx: usize = bracket_content.parse().map_err(|_| {
+//                         OutputPluginError::OutputPluginFailed(format!(
+//                             "invalid bracket segment '[{bracket_content}]' in path '{path}': \
+//                             expected an integer index or a quoted string key"
+//                         ))
+//                     })?;
+//                     PathSegment::Index(idx)
+//                 };
+//                 segments.push(segment);
+//             }
+//             other => current_key.push(other),
+//         }
+//     }
 
-    if !current_key.is_empty() {
-        segments.push(PathSegment::Key(current_key));
-    }
+//     if !current_key.is_empty() {
+//         segments.push(PathSegment::Key(current_key));
+//     }
 
-    Ok(segments)
-}
-
-/// Recursively walk `root` along `segments` and write `value` at the final location.
-/// Intermediate objects that do not exist are created automatically.
-pub fn set_path(
-    root: &mut serde_json::Value,
-    segments: &[PathSegment],
-    value: serde_json::Value,
-) -> Result<(), OutputPluginError> {
-    match segments {
-        [] => Err(OutputPluginError::OutputPluginFailed(
-            "empty output path".to_string(),
-        )),
-
-        [PathSegment::Key(k)] => {
-            root.as_object_mut()
-                .ok_or_else(|| {
-                    OutputPluginError::OutputPluginFailed(format!(
-                        "cannot write key '{k}' into a non-object JSON value"
-                    ))
-                })?
-                .insert(k.clone(), value);
-            Ok(())
-        }
-
-        [PathSegment::Index(i)] => {
-            let arr = root.as_array_mut().ok_or_else(|| {
-                OutputPluginError::OutputPluginFailed(format!(
-                    "cannot index a non-array JSON value with [{i}]"
-                ))
-            })?;
-            if *i < arr.len() {
-                arr[*i] = value;
-                Ok(())
-            } else {
-                Err(OutputPluginError::OutputPluginFailed(format!(
-                    "array index {i} is out of bounds (length {})",
-                    arr.len()
-                )))
-            }
-        }
-
-        [PathSegment::Key(k), rest @ ..] => {
-            if !root.is_object() {
-                *root = serde_json::Value::Object(serde_json::Map::new());
-            }
-            let child = root
-                .as_object_mut()
-                .unwrap()
-                .entry(k.clone())
-                .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
-            set_path(child, rest, value)
-        }
-
-        [PathSegment::Index(i), rest @ ..] => {
-            let arr = root.as_array_mut().ok_or_else(|| {
-                OutputPluginError::OutputPluginFailed(format!(
-                    "cannot index a non-array JSON value with [{i}]"
-                ))
-            })?;
-            let child = arr.get_mut(*i).ok_or_else(|| {
-                OutputPluginError::OutputPluginFailed(format!("array index {i} is out of bounds"))
-            })?;
-            set_path(child, rest, value)
-        }
-    }
-}
+//     Ok(segments)
+// }
 
 /// Append an error entry `{ "expr": ..., "error": ... }` to the array located
 /// at `segments` within `root`. If the target location does not yet hold an
@@ -429,8 +366,144 @@ pub fn eval_and_write(
     )
 }
 
+/// Recursively walk `root` along `segments` and write `value` at the final location.
+/// Intermediate objects that do not exist are created automatically.
+pub fn set_path(
+    root: &mut serde_json::Value,
+    segments: &[PathSegment],
+    value: serde_json::Value,
+) -> Result<(), OutputPluginError> {
+    match segments {
+        [] => Err(OutputPluginError::OutputPluginFailed(
+            "empty output path".to_string(),
+        )),
+
+        [PathSegment::Key(k)] => {
+            root.as_object_mut()
+                .ok_or_else(|| {
+                    OutputPluginError::OutputPluginFailed(format!(
+                        "cannot write key '{k}' into a non-object JSON value"
+                    ))
+                })?
+                .insert(k.clone(), value);
+            Ok(())
+        }
+
+        [PathSegment::Index(i)] => {
+            let arr = root.as_array_mut().ok_or_else(|| {
+                OutputPluginError::OutputPluginFailed(format!(
+                    "cannot index a non-array JSON value with [{i}]"
+                ))
+            })?;
+            if *i < arr.len() {
+                arr[*i] = value;
+                Ok(())
+            } else {
+                Err(OutputPluginError::OutputPluginFailed(format!(
+                    "array index {i} is out of bounds (length {})",
+                    arr.len()
+                )))
+            }
+        }
+
+        [PathSegment::Key(k), rest @ ..] => {
+            if !root.is_object() {
+                *root = serde_json::Value::Object(serde_json::Map::new());
+            }
+            let child = root
+                .as_object_mut()
+                .unwrap()
+                .entry(k.clone())
+                .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
+            set_path(child, rest, value)
+        }
+
+        [PathSegment::Index(i), rest @ ..] => {
+            let arr = root.as_array_mut().ok_or_else(|| {
+                OutputPluginError::OutputPluginFailed(format!(
+                    "cannot index a non-array JSON value with [{i}]"
+                ))
+            })?;
+            let child = arr.get_mut(*i).ok_or_else(|| {
+                OutputPluginError::OutputPluginFailed(format!("array index {i} is out of bounds"))
+            })?;
+            set_path(child, rest, value)
+        }
+    }
+}
+
+/// Parse a JSONPath-style string into segments for writing.
+///
+/// Supports dot notation (`$.a.b`) and bracket array indices (`$.a[0].b`).
+/// The leading `$.` or `$` prefix is stripped before parsing.
+pub fn parse_path_segments(path: &str) -> Result<Vec<PathSegment>, OutputPluginError> {
+    let stripped = path
+        .strip_prefix("$.")
+        .or_else(|| path.strip_prefix('$'))
+        .unwrap_or(path);
+
+    if stripped.is_empty() {
+        return Err(OutputPluginError::OutputPluginFailed(
+            "output path must not be empty after '$'".to_string(),
+        ));
+    }
+
+    let mut segments = Vec::new();
+    let mut current_key = String::new();
+    let mut chars = stripped.chars().peekable();
+
+    while let Some(c) = chars.next() {
+        match c {
+            '.' => {
+                if !current_key.is_empty() {
+                    segments.push(PathSegment::Key(std::mem::take(&mut current_key)));
+                }
+            }
+            '[' => {
+                if !current_key.is_empty() {
+                    segments.push(PathSegment::Key(std::mem::take(&mut current_key)));
+                }
+                let mut bracket_content = String::new();
+                for ic in chars.by_ref() {
+                    if ic == ']' {
+                        break;
+                    }
+                    bracket_content.push(ic);
+                }
+                // Quoted string key: ['foo'] or ["foo"]
+                let segment = if (bracket_content.starts_with('\'')
+                    && bracket_content.ends_with('\''))
+                    || (bracket_content.starts_with('"') && bracket_content.ends_with('"'))
+                {
+                    let key = bracket_content[1..bracket_content.len() - 1].to_string();
+                    PathSegment::Key(key)
+                } else {
+                    // Numeric array index: [0]
+                    let idx: usize = bracket_content.parse().map_err(|_| {
+                        OutputPluginError::OutputPluginFailed(format!(
+                            "invalid bracket segment '[{bracket_content}]' in path '{path}': \
+                            expected an integer index or a quoted string key"
+                        ))
+                    })?;
+                    PathSegment::Index(idx)
+                };
+                segments.push(segment);
+            }
+            other => current_key.push(other),
+        }
+    }
+
+    if !current_key.is_empty() {
+        segments.push(PathSegment::Key(current_key));
+    }
+
+    Ok(segments)
+}
+
 #[cfg(test)]
 mod tests {
+    use crate::plugin::output::default::eval::ops::{record_error, set_path};
+
     use super::*;
     use serde_json::json;
